[refactor] create subdirectories for each page, add linkedin and github links to nav
