+++ public/main/components/footer.jsx
